This is what i have for android so far. Havent had time to replace your production plugin with what i made here in my own app so this code is UNTESTED.
Just doing a commit as a backup. and maybe you see some obvious things.

So far added UserCategories and UserActions.
Removed some bits from the other library as they have a way to customize what Activities you start when you click an action. that seemed a bit overkill for a first version.